Write a code generator for the Theme object
Previously, the Theme object parsed the theme data at run-time
and populated a QQmlPropertyMap with the values.

This meant that any bindings or expressions which resolved
  Theme.geometry.some.value
or similar, could not be optimized by the compiler.

This commit adds a code generator which ingests the theme .json
files and produces a strongly-typed theme object hierarchy.