Locate fonts that can render Arabic and Chinese scripts
Arabic characters can be rendered by DejaVuSans, and Chinese
characters can be rendered by DroidSansFallback.ttf. Try to locate
these on device or from wasm builds. If they cannot be found, fall
back to the default MuseoSans font.

Use a URL instead of a file path for the LanguageModel model data so
that the file path is resolved correctly when setting the source for
a QML FontLoader.