Avoid jittery gauge effects when Brief gauge layout changes
In _leftGaugeParameters() and _rightGaugeParameters(), fix the gauge
count and height at the start of the function. Otherwise, the values
may change throughout the function when the property bindings update
themselves, leading to incorrect values in the returned paramters,
which need to be corrected later on. This was creating jittery gauge
updates as the layout was quickly changed multiple times within a
short time.