Overview: don't force connectors to be in a straight line
This ability to straighten a connector only worked for the AC input
box, and this was causing a bug where the alternator widget would not
show a widget connector if it was the only left-hand widget.

Remove the feature as it doesn't work correctly in this regard, and
also causes the connector anchor to jump around unnecessarily when the
widget layout changes. The connector will be straight anyway if two
horizontally adjacent boxes have the same height.